Update filter-feeds.py
Modified to download the newest copy of a blacklist hosted on the website. It can be commented out to simply open and use a local list of blacklisted domains. 

news-admin.php appends selected domains to the blacklist in the folder it resides in. This script downloads that blacklist, allowing for an easier method than hand entering domains.

Retrieving the blacklist also logs the action to feedlog.txt